% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylog.R
\name{mylog}
\alias{mylog}
\title{My logarithm function}
\usage{
mylog(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
a vector of the same length
as \code{x} with the \code{log} of
\code{x}.
}
\description{
This is a function that
computes the logarithm of a vector
}
\examples{
log(1)
log(2)

}
